# docs/04-api/openapi.yaml
openapi: 3.0.3
info:
  title: RAWAQA E-Commerce API
  description: |
    REST API specification for RAWAQA custom e-commerce platform.
    **Status:** Skeleton — API Not Yet Implemented.
    Based on current frontend prototype (`js/main.js` PRODUCTS) and commercial scope.
  version: 1.0.0
  contact:
    name: RAWAQA Development Team

servers:
  - url: http://localhost:3000/api/v1
    description: Development
  - url: https://api.rawaqa.example.com/v1
    description: Production (placeholder domain)

tags:
  - name: Auth
  - name: Products
  - name: Cart
  - name: Orders
  - name: Admin

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already exists

  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  /auth/me:
    get:
      tags: [Auth]
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [Relax, Game, Kids, Outdoor]
        - name: min_price
          in: query
          schema:
            type: number
        - name: max_price
          in: query
          schema:
            type: number
        - name: in_stock
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [recommended, price_asc, price_desc, newest]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'

  /categories:
    get:
      tags: [Products]
      summary: List categories
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'

  /cart:
    get:
      tags: [Cart]
      summary: Get cart
      security:
        - bearerAuth: []
        - sessionAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    delete:
      tags: [Cart]
      summary: Clear cart
      security:
        - bearerAuth: []
        - sessionAuth: []
      responses:
        '204':
          description: Cleared

  /cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart
      security:
        - bearerAuth: []
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '409':
          description: Insufficient stock

  /cart/items/{itemId}:
    patch:
      tags: [Cart]
      summary: Update cart item quantity
      security:
        - bearerAuth: []
        - sessionAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    delete:
      tags: [Cart]
      summary: Remove cart item
      security:
        - bearerAuth: []
        - sessionAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed

  /orders:
    post:
      tags: [Orders]
      summary: Create order (checkout)
      security:
        - bearerAuth: []
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
        '422':
          description: Empty cart or business rule violation
    get:
      tags: [Orders]
      summary: Customer order history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'

  /orders/{id}:
    get:
      tags: [Orders]
      summary: Order detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /orders/track/{orderNumber}:
    get:
      tags: [Orders]
      summary: Track order by public number
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
            example: RWQ-10482
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTracking'
        '404':
          $ref: '#/components/responses/NotFound'

  /admin/dashboard:
    get:
      tags: [Admin]
      summary: Dashboard KPIs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboard'

  /admin/products:
    get:
      tags: [Admin]
      summary: List all products (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      tags: [Admin]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Created

  /admin/products/{id}:
    patch:
      tags: [Admin]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: OK
    delete:
      tags: [Admin]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  /admin/orders:
    get:
      tags: [Admin]
      summary: List orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK

  /admin/orders/{id}/status:
    patch:
      tags: [Admin]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, preparing, shipped, delivered, cancelled]
                note:
                  type: string
      responses:
        '200':
          description: OK

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionAuth:
      type: apiKey
      in: cookie
      name: session_id

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object

    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        phone:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [customer, admin]

    RegisterRequest:
      type: object
      required: [email, phone, password, full_name]
      properties:
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        full_name:
          type: string

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string

    ProductVariant:
      type: object
      properties:
        id:
          type: string
        color:
          type: string
        size:
          type: string
        sku:
          type: string
        price:
          type: number
        stock:
          type: integer

    Product:
      type: object
      properties:
        id:
          type: string
          example: cloud-lounger
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
          default: EGP
        category:
          type: string
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              alt:
                type: string
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariant'
        in_stock:
          type: boolean
        featured:
          type: boolean

    ProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category_id:
          type: string
        price:
          type: number
        featured:
          type: boolean
        active:
          type: boolean

    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
        variant_id:
          type: string
        name:
          type: string
        variant_label:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
        line_total:
          type: number

    Cart:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
        item_count:
          type: integer

    AddCartItemRequest:
      type: object
      required: [product_id, variant_id, quantity]
      properties:
        product_id:
          type: string
        variant_id:
          type: string
        quantity:
          type: integer
          minimum: 1

    CreateOrderRequest:
      type: object
      required: [customer, shipping_address, payment_method]
      properties:
        customer:
          type: object
          required: [full_name, phone]
          properties:
            full_name:
              type: string
            phone:
              type: string
            email:
              type: string
        shipping_address:
          type: object
          required: [governorate, city, street]
          properties:
            governorate:
              type: string
            city:
              type: string
            street:
              type: string
            notes:
              type: string
        payment_method:
          type: string
          enum: [cod]
        cart_id:
          type: string

    OrderItem:
      type: object
      properties:
        product_id:
          type: string
        product_name:
          type: string
        variant_label:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
        line_total:
          type: number

    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
          example: RWQ-10482
        status:
          type: string
          enum: [pending, confirmed, preparing, shipped, delivered, cancelled]
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        currency:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        customer:
          type: object
        shipping_address:
          type: object
        odoo_sync_status:
          type: string
          enum: [pending, synced, failed]
        sms_status:
          type: string
          enum: [pending, sent, failed]
        created_at:
          type: string
          format: date-time

    OrderTracking:
      type: object
      properties:
        order_number:
          type: string
        status:
          type: string
        timeline:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              timestamp:
                type: string
              completed:
                type: boolean
        tracking_number:
          type: string
        estimated_delivery:
          type: string

    AdminDashboard:
      type: object
      properties:
        orders_today:
          type: integer
        orders_pending:
          type: integer
        products_count:
          type: integer
        recent_orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'